requirements.txt
flask
pylightning
Flask-BasicAuth
